requirements.txt
telethon>=1.34.0
sqlalchemy>=2.0
aiosqlite>=0.19
python-dotenv>=1.0
pandas
openpyxl